refactor TaskQueueZkImpl (#1591)
* fix #1515

* sleep when resource in not satisfy. fix #1522

* add sleep 1s for no command

* fix MasterBaseTaskExecThread submit method bug

* updates

* add log

* delete lombok

* remove duplicate code

* refactor TaskQueueZkImpl

* ignore First , we have to rewrite

* updates
