+++ dolphinscheduler-common/src/test/java/org/apache/dolphinscheduler/common/queue/TaskQueueZKImplTest.java
@@ -32,10 +32,8 @@ import static org.junit.Assert.*;
 /**
  * task queue test
  */
+@Ignore
 public class TaskQueueZKImplTest extends BaseTaskQueueTest  {
-
-
-
 
     @Before
     public void before(){
